requirements.txt
flask==3.0.3
flask-limiter==3.7.0
google-generativeai==0.7.2
google-cloud-firestore==2.16.1
python-dotenv==1.0.1
gunicorn==22.0.0
pytest==8.1.1
pytest-flask==1.3.0
